+++ client/src/components/embedding/index.js
@@ -1,32 +1,93 @@
 import React from "react";
 import { connect } from "react-redux";
+import {
+  ButtonGroup,
+  Popover,
+  Button,
+  Radio,
+  RadioGroup,
+  Tooltip,
+  Position,
+} from "@blueprintjs/core";
+import * as globals from "../../globals";
+import actions from "../../actions";
 
 @connect((state) => ({
   layoutChoice: state.layoutChoice,
+  schema: state.annoMatrix?.schema,
+  crossfilter: state.obsCrossfilter,
 }))
-class Embedding extends React.Component {
+class Embedding extends React.PureComponent {
   constructor(props) {
     super(props);
     this.state = {};
   }
 
+  handleLayoutChoiceChange = (e) => {
+    const { dispatch } = this.props;
+    dispatch(actions.layoutChoiceAction(e.currentTarget.value));
+  };
+
   render() {
-    const { layoutChoice } = this.props;
+    const { layoutChoice, schema, crossfilter } = this.props;
     return (
-      <div
-        id="embedding"
-        data-testclass="embedding"
+      <ButtonGroup
         style={{
           position: "absolute",
           display: "inherit",
           left: 8,
           bottom: 8,
-          zIndex: 1,
-          fontWeight: 700,
+          zIndex: 9999,
         }}
       >
-        {layoutChoice?.current}
-      </div>
+        <Popover
+          target={
+            <Tooltip
+              content="Select embedding for visualization"
+              position="top"
+              hoverOpenDelay={globals.tooltipHoverOpenDelay}
+            >
+              <Button
+                type="button"
+                data-testid="layout-choice"
+                icon="heatmap"
+                minimal
+                id="embedding"
+                style={{
+                  cursor: "pointer",
+                }}
+              >
+                {layoutChoice?.current}: {crossfilter.countSelected()} out of{" "}
+                {schema?.dataframe?.nObs} total cells
+              </Button>
+            </Tooltip>
+          }
+          minimal /* removes arrow */
+          position={Position.TOP_LEFT}
+          content={
+            <div
+              style={{
+                display: "flex",
+                justifyContent: "flex-start",
+                alignItems: "flex-start",
+                flexDirection: "column",
+                padding: 10,
+                width: 400,
+              }}
+            >
+              <RadioGroup
+                label="Embedding Choice"
+                onChange={this.handleLayoutChoiceChange}
+                selectedValue={layoutChoice.current}
+              >
+                {layoutChoice.available.map((name) => (
+                  <Radio label={name} value={name} key={name} />
+                ))}
+              </RadioGroup>
+            </div>
+          }
+        />
+      </ButtonGroup>
     );
   }
 }
